paradigmctf.py/ctf_solvers/solver.py
import os

from pwn import *
import yaml


class TicketedRemote:
    def __enter__(self) -> remote:
        challenge_id = sys.path[0].split(os.path.sep)[-2]

        env = os.getenv("ENV", "local")
        if env == "local":
            host = "127.0.0.1"
        elif env == "dev":
            host = f"{challenge_id}.dev.ctf.paradigm.xyz"
        elif env == "prod":
            host = f"{challenge_id}.challenges.paradigm.xyz"
        else:
            raise Exception("unsupported env")

        self.__r = remote(host, "1337")

        data = self.__r.recvuntil(b"?")
        if "ticket" not in data.decode():
            self.__r.unrecv(data)
            return self.__r

        if env == "dev":
            ticket = "dev2023"
        else:
            ticket = os.getenv("SECRET") + ":healthcheck-team:" + challenge_id

        self.__r.send(ticket.encode("utf8"))
        self.__r.send(b"\n")

        return self.__r

    def __exit__(self, exc_type, exc_val, exc_tb):
        self.__r.close()


def kill_instance():
    with TicketedRemote() as r:
        r.recvuntil(b"?")
        r.send(b"2\n")
        r.recvall()


def launch_instance():
    with TicketedRemote() as r:
        r.recvuntil(b"?")
        r.send(b"1\n")
        try:
            r.recvuntil(b"---\n")
        except:
            log = r.recvall().decode("utf8")
            print(log)
            raise Exception("failed to create instance")

        data_raw = r.recvall().decode("utf8")

    # todo: this fails when the private key has a leading zero
    return yaml.safe_load(data_raw)
