paradigmctf.py/requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
bcrypt==4.0.1
bitarray==2.8.2
cachetools==5.3.2
capstone==5.0.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colored-traceback==0.3.0
cryptography==41.0.5
cytoolz==0.12.2
docker==6.1.3
eth-abi==4.2.1
eth-account==0.10.0
eth-hash==0.5.2
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.5.2
eth-utils==2.3.1
fastapi==0.104.1
frozenlist==1.4.0
google-auth==2.23.4
h11==0.14.0
hexbytes==0.3.1
idna==3.4
intervaltree==3.1.0
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
kubernetes==28.1.0
lru-dict==1.2.0
Mako==1.3.0
MarkupSafe==2.1.3
multidict==6.0.4
oauthlib==3.2.2
packaging==23.2
paramiko==3.3.1
parsimonious==0.9.0
plumbum==1.8.2
protobuf==4.25.0
psutil==5.9.6
pwntools==4.11.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.19.0
pydantic==2.4.2
pydantic_core==2.10.1
pyelftools==0.30
Pygments==2.16.1
PyNaCl==1.5.0
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.8.2
pyunormalize==15.0.0
PyYAML==6.0.1
redis==5.0.1
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rlp==3.0.0
ROPGadget==7.4
rpds-py==0.12.0
rpyc==5.3.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
starlette==0.27.0
toolz==0.12.0
typing_extensions==4.8.0
unicorn==2.0.1.post1
urllib3==1.26.18
uvicorn==0.24.0.post1
web3==6.11.3
websocket-client==1.6.4
websockets==12.0
yarl==1.9.2